+++ Game/Modules/Burning/Components/BurningTimerComponent.cs
@@ -8,15 +8,15 @@ namespace Game.Modules.Burning.Components;
 /// 0-to-1 stack transition, removed once stacks reach 0 (see BurningSystem). Countdown to the
 /// next damage tick; gaining an additional stack while already burning must not reset it.
 /// </summary>
-public struct BurningTimerComponent(int framesUntilNextTick, int stackCount) : ITickCountdown, IStatusEffectStackCount
+public struct BurningTimerComponent(ushort framesUntilNextTick, byte stackCount) : ITickCountdown, IStatusEffectStackCount
 {
-    public int FramesUntilNextTick { get; set; } = framesUntilNextTick;
+    public ushort FramesUntilNextTick { get; set; } = framesUntilNextTick;
 
     /// <summary>
     /// Cached running total of this entity's Burning stacks, kept in sync by
     /// BurningEffects.ApplyStack (increment) and BurningSystem.Tick (decrement)
     /// </summary>
-    public int StackCount { get; set; } = stackCount;
+    public byte StackCount { get; set; } = stackCount;
 
     public override readonly string ToString() => $"FramesUntilNextTick : {FramesUntilNextTick}\nStackCount : {StackCount}";
 }
